refactor lock_utils.py to utilities/lock_utils.py
