+++ ulc_mm_package/hardware/real/motorcontroller.py
@@ -13,7 +13,7 @@ import functools
 import threading
 import pigpio
 
-from ulc_mm_package.lock_utils import lock_no_block
+from ulc_mm_package.utilities.lock_utils import lock_no_block
 from ulc_mm_package.hardware.hardware_constants import (
     FULL_STEP_TO_TRAVEL_DIST_UM,
     DEFAULT_FULL_STEP_HOMING_TIMEOUT,
