+++ ulc_mm_package/hardware/sim/motorcontroller.py
@@ -3,7 +3,7 @@ import time
 import functools
 import threading
 
-from ulc_mm_package.lock_utils import lock_no_block
+from ulc_mm_package.utilities.lock_utils import lock_no_block
 from ulc_mm_package.hardware.hardware_constants import (
     FULL_STEP_TO_TRAVEL_DIST_UM,
     DEFAULT_FULL_STEP_HOMING_TIMEOUT,
