+++ ulc_mm_package/neural_nets/YOGOInference.py
@@ -8,7 +8,7 @@ from collections import namedtuple
 from typing import Any, List, Union
 
 
-from ulc_mm_package.lock_utils import lock_timeout
+from ulc_mm_package.utilities.lock_utils import lock_timeout
 from ulc_mm_package.scope_constants import CameraOptions, CAMERA_SELECTION
 from ulc_mm_package.neural_nets.NCSModel import (
     NCSModel,
